Support for writing H5GF archives from outside ALPScore.
Allows to create archives satisfying the ALPScore H5GF specification
from outside ALPScore.  The rationale for this is that it is useful
in interfacing other codes from and to ALPScore applications and
allowing alpscore to be used 'interactively' from a Python shell by
preparing the input files manually.

This in a sense closes ALPScore issue #254.